requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
openpyxl==3.1.5
cairosvg==2.7.1
PyYAML==6.0.2
httpx==0.27.2
pypdf==5.3.1
